// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          artist_name_name: string | null
          bio: string | null
          birth_date: string | null
          chave_pix: string | null
          email: string | null
          full_name: string | null
          instagram_handle: string | null
          location: string | null
          music_links: string | null
          name: string | null
          phone: string | null
          portfolio_url: string | null
          presskit_url: string | null
          role: string | null 
          soundcloud_url: string | null
          user_id: string
          youtube_url: string | null
        }
        Insert: {
          artist_name_name?: string | null
          bio?: string | null
          birth_date?: string | null
          chave_pix?: string | null
          email?: string | null
          full_name?: string | null
          instagram_handle?: string | null
          location?: string | null
          music_links?: string | null
          name?: string | null
          phone?: string | null
          portfolio_url?: string | null
          presskit_url?: string | null
          role?: string
          soundcloud_url?: string | null
          user_id: string
          youtube_url?: string | null
        }
        Update: {
          artist_name_name?: string | null
          bio?: string | null
          birth_date?: string | null
          chave_pix?: string | null
          email?: string | null
          full_name?: string | null
          instagram_handle?: string | null
          location?: string | null
          music_links?: string | null
          name?: string | null
          phone?: string | null
          portfolio_url?: string | null
          presskit_url?: string | null
          role?: string
          soundcloud_url?: string | null
          user_id?: string
          youtube_url?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}